chore(tests): extract trim-end semantics tests to their own module
Fulfills the maintainer promise on PR #493: the new TestTrimEndSemantics
block moves to tests/test_engine_trim_end.py, bringing test_engine.py
back under the 800-LOC module ceiling (938 -> 867); authorship preserved
(WohaibHasan, PR #493) per the contributor policy.
